Trivial PID controller, but same behavior as previous demo.
This implementation assumes that the output of the PID controller is
directly suitable as a canard angle, which ignores a fair bit of
velocity-dependent physics modelling that we may want.

It then preserves the previous demo's behavior by setting the
controller's coefficients to (1, 0, 0), which we'll want to tune. A lot.