Made roll launchdetect depend on arm, fixed not sending servo disable packets
The servo enable and disable messages to the FC are for debug, and should
cause rollcontrol to send packets to the servos reguardless of launch
or arm state. Rollcontrol now only looks for launch while armed.

The sending disable packets should ideally be just in set_servo_enable()
but since this is called in an init function there is no guarantee that
the servo socket is open. Instead as a cheap hack, set_canard_angle(0) is
placed wherever the servo should be disabled, after the servo disable flag
is set.